requirements.txt
# AD-DRS: Adaptive Merging in Drift-Resistant Space
# Requirements file for the enhanced LoRA-Sub-DRS implementation

# Core PyTorch and deep learning
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2

# Vision Transformer and model utilities
timm==0.6.7

# Scientific computing and numerical analysis
numpy>=1.21.0
scipy>=1.7.0

# Visualization and plotting
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.64.0

# Image processing
Pillow>=9.0.0

# Data handling and serialization
pandas>=1.3.0

# Logging and configuration
pyyaml>=6.0

# Optional: For enhanced analysis and visualization
plotly>=5.0.0
ipywidgets>=7.6.0

# Optional: For Jupyter notebook support (if using notebooks)
jupyter>=1.0.0
ipykernel>=6.0.0

# Optional: For distributed training (if needed)
# accelerate>=0.12.0

# Development and testing (optional)
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0

# Memory profiling (optional, for debugging)
psutil>=5.8.0

# For reproducibility
setuptools>=60.0.0
wheel>=0.37.0

ipdb==0.13.13

scikit-learn==1.7.1
